Fix header guard names; Move ResolvedAddress
