+++ BUILD
@@ -133,7 +133,6 @@ GRPC_PUBLIC_EVENT_ENGINE_HDRS = [
     "include/grpc/event_engine/event_engine.h",
     "include/grpc/event_engine/port.h",
     "include/grpc/event_engine/slice_allocator.h",
-    "include/grpc/event_engine/sockaddr.h",
 ]
 
 GRPC_SECURE_PUBLIC_HDRS = [
